+++ client/src/components/about/About.js
@@ -3,7 +3,7 @@ import SVGImage from 'react-native-svg-image';
 
 import { Platform, StyleSheet, View, Text, ScrollView, Image } from 'react-native';
 import SessionBar from '../navbar/SessionBar';
-// import './About.css';
+import styles from './AboutCss';
 
 const About = (props) => (
   <View style={styles.main}>
@@ -110,112 +110,3 @@ const About = (props) => (
 );
 
 export default About;
-
-const styles = StyleSheet.create({
-  ...Platform.select({
-    ios: {
-      main: {
-        backgroundColor: '#333',
-        flex: 1,
-      },
-      view: {
-        paddingLeft: 20,
-        paddingRight: 30,
-        paddingTop: 30,
-      },
-      text: {
-        textAlign: 'center',
-        fontSize: 30,
-        color: '#D7D7D7',
-      },
-      underlineText: {
-        textAlign: 'center',
-        fontStyle: 'italic',
-        fontSize: 30,
-        color: '#D7D7D7',
-      },
-      contributor: {
-        borderStyle: 'solid',
-        borderWidth: 5,
-        borderColor: 'white',
-        padding: 15,
-        margin: 5,
-        flex: 1,
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
-        shadowOffset: { width: 2, height: 2 },
-        shadowColor: 'white',
-        shadowOpacity: 1.0,
-      },
-      techTitle: {
-        backgroundColor: 'midnightblue',
-        textAlign: 'center',
-        color: 'white',
-        fontSize: 30,
-        textShadowColor: 'black',
-        textShadowOffset: { width: 5, height: 5 },
-        textShadowRadius: 4,
-      },
-      techHeader: {
-        backgroundColor: 'midnightblue',
-        marginTop: 50,
-      },
-      techRow: {
-        backgroundColor: 'midnightblue',
-        padding: 15,
-        flex: 1,
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-      },
-      techRow2: {
-        backgroundColor: 'midnightblue',
-        padding: 15,
-        flex: 1,
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-      },
-      image: {
-        width: 300,
-        height: 300,
-      },
-      svg: {
-        margin: 5,
-        width: 50,
-        height: 50,
-      },
-    },
-    android: {
-      main: {
-        backgroundColor: '#333',
-        flex: 1,
-      },
-      view: {
-        paddingLeft: 20,
-        paddingRight: 30,
-        paddingTop: 30,
-      },
-      text: {
-        textAlign: 'center',
-        fontSize: 30,
-        color: '#D7D7D7',
-      },
-      contributor: {
-        margin: 'auto',
-        borderStyle: 'solid',
-        borderWidth: 5,
-        borderColor: 'white',
-        padding: 15,
-        // boxShadow: 2px 2px 10px white,
-        lineHeight: 1.5,
-      },
-      image: {
-        width: 300,
-        height: 300,
-      },
-    }
-  })
-
-})
